pyproject.toml
[tool.poetry]
name="CreateReport"
version="0.1.0"
description=""

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-restful = "^0.3.10"
flask-cors = "^5.0.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
flask-jwt-extended = "^4.6.0"
flask-marshmallow = "^1.2.1"
requests = "2.32.3"
loguru = "^0.7.2"
pyyaml = "^6.0.2"
python-docx = "1.2.0"
docxtpl = "^0.20.1"
openpyxl = "3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
